requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
click==8.1.7
questionary==2.0.1
tabulate==0.9.0
flask==3.0.0
flask-limiter==3.5.0
pytest==7.4.3
pytest-cov==4.1.0
